* origin/topic/robin/gh-574-hook-priority:
  Update tests.
  Add documentation for hook priorities.
  Add tests for Spicy-side priorities.
  Add HILTI tests for hook priorities.
  Call hook implementations in order of priority.
  Pass hook priorities on to linker.
  Add HILTI-side validation for hook priorities.
  Add helper to retrieve static integer value from attribute.

+++ hilti/toolchain/include/compiler/detail/cxx/unit.h
@@ -49,7 +49,7 @@ struct Join {
     cxx::declaration::Function callee; /**< callee function to execute through linker function */
     std::list<cxx::declaration::Type>
         aux_types;    /**< additional types the linker needs to declare for external prototype to work */
-    int priority = 0; /**< Priority determing the order between callees; higher prioeriy calleed will be called first */
+    int64_t priority = 0; /**< Priority determining the order between callees; higher priority callees will be called first */
     bool declare_only = false; /**< only declare the joined C++ function, don't generate the implementation */
 
     bool operator<(const Join& other) const {
